Fix: could not be closed twice, gave NullPointerException.

+++ adaptors/CommandlineSsh/src/org/gridlab/gat/io/cpi/copyingFileInputStream/CopyingFileInputStreamAdaptor.java
@@ -114,8 +114,10 @@ public class CopyingFileInputStreamAdaptor extends FileInputStreamCpi {
         } catch (IOException e) {
             // ignore
         }
-
-        localFile.delete();
+        if (localFile != null) {
+            // If null, already closed.
+            localFile.delete();
+        }
         localFile = null;
         in = null;
     }
